+++ Userland/Applications/FileManager/FileUtils.h
@@ -11,7 +11,7 @@
 #include <LibGUI/Forward.h>
 #include <sys/stat.h>
 
-namespace FileUtils {
+namespace FileManager {
 
 enum class FileOperation {
     Copy = 0,
@@ -20,4 +20,6 @@ enum class FileOperation {
 
 void delete_path(String const&, GUI::Window*);
 void delete_paths(Vector<String> const&, bool should_confirm, GUI::Window*);
+
+void run_file_operation([[maybe_unused]] FileOperation operation, Vector<String> const& sources, String const& destination, GUI::Window* parent_window);
 }
